// src/generated/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CustomerType = {
  DEALER: 'DEALER',
  AUTHORIZED_SERVICE: 'AUTHORIZED_SERVICE',
  ACTIVE: 'ACTIVE',
  POTENTIAL: 'POTENTIAL',
  CORPORATE_END_USER: 'CORPORATE_END_USER',
  PROJECT_CUSTOMER: 'PROJECT_CUSTOMER',
  INDIVIDUAL: 'INDIVIDUAL',
  EXPORT_DISTRIBUTOR: 'EXPORT_DISTRIBUTOR',
  EXPORT_PROJECT: 'EXPORT_PROJECT'
} as const

export type CustomerType = (typeof CustomerType)[keyof typeof CustomerType]


export const ActivityType = {
  INTRO_CALL: 'INTRO_CALL',
  WHATSAPP_MESSAGE: 'WHATSAPP_MESSAGE',
  EMAIL: 'EMAIL',
  QUOTE_SENT: 'QUOTE_SENT',
  SHIPMENT: 'SHIPMENT',
  INSTALLATION: 'INSTALLATION',
  NOTE: 'NOTE'
} as const

export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType]